Parallel computation of map

+++ lib/src/main/java/formatter/SequenceParserResult.java
@@ -4,6 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class SequenceParserResult {
-    public double[] sequence;
+    public float[] sequence;
     public List<String> errors = new ArrayList<>();
 }
