+++ lib/src/main/java/provider/SequenceProvider.java
@@ -1,5 +1,5 @@
 package provider;
 
 public interface SequenceProvider {
-    double[] getSequenceByName(String sequenceName);
+    float[] getSequenceByName(String sequenceName);
 }
